pyproject.toml
[tool.poetry]
name = "djangosvelte"
version = "0.1.0"
description = "Django Svelte Project"
authors = ["grischan <grischan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
whitenoise = "^6.2.0"
django-rest-framework = "^0.1.0"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
